backend/requirements.txt

Flask==2.2.3
Flask-Cors==3.0.10
Pillow==9.5.0
requests==2.28.2
python-dotenv==1.0.0
